==> MVMR/OSA_and_BMI_exposure_to_T2D_outcome_BMI_UNadj_GWASs/2_filter_SNPs_Exposure_OSA_and_BMI.R <==
library(readr)
library(dplyr)
library(data.table)

#read Exposures data
#This OSA file was the outcome when using T2D and BMI as exposures
OSA <- fread("/Users/yana/Library/CloudStorage/OneDrive-BethIsraelLaheyHealth/2022_diabetes_PRS_interaction_sleep/Data/post_processed_data_for_MR_analysyis/T2D_BMI_UNadj_to_OSA_BMI_UNadj/Outcome_OSA_BMI_UNadj_distinct_SNPs_sub.csv")
BMI <- fread("/Users/yana/Library/CloudStorage/OneDrive-BethIsraelLaheyHealth/2022_diabetes_PRS_interaction_sleep/Data/post_processed_data_for_MR_analysyis/OSA_and_BMI_exposure_to_T2D_outcome_BMI_UNadj_GWASs/Exposure_BMI_for_extract_exposure.csv")


#Match SNPs between TWO Exposures to make sure that the SNPs are present in both
matched_SNPs_OSA_BMI <- intersect(OSA$rsID, BMI$rsID)

#subset each Exposure to the SNPs that matched
OSA_GWAS_sub <- OSA[which(OSA$rsID %in% matched_SNPs_OSA_BMI),]
BMI_GWAS_sub <- BMI[which(BMI$rsID %in% matched_SNPs_OSA_BMI),]

#filter variants to only include the ones that pass the p-value threshold of p < 5*10^-8
# p < 5*10^-8 DONE
# p < 10^-7 
# p < 10^-5 

#threshold <- "5_times_10_to_negative_8"
#threshold <- "10_to_negative_7"
threshold <- "10_to_negative_5"



#select instruments from each Exposure GWAS
sig_OSA <- OSA_GWAS_sub[which(as.numeric(OSA_GWAS_sub$PValue) < 10^-5),]
sig_BMI <- BMI_GWAS_sub[which(as.numeric(BMI_GWAS_sub$Pvalue) < 10^-5),]

#save filtered SNPs of the Exposure GWASs
write.csv(sig_OSA, paste0("/Users/yana/Library/CloudStorage/OneDrive-BethIsraelLaheyHealth/2022_diabetes_PRS_interaction_sleep/Data/post_processed_data_for_MR_analysyis/OSA_and_BMI_exposure_to_T2D_outcome_BMI_UNadj_GWASs/SNPs_filtered_by_p/",threshold,"/Exposure_OSA_BMI_UNadj_distinct_SNPs_filtered_by_p_",threshold,".csv"), row.names = FALSE)

write.csv(sig_BMI, paste0("/Users/yana/Library/CloudStorage/OneDrive-BethIsraelLaheyHealth/2022_diabetes_PRS_interaction_sleep/Data/post_processed_data_for_MR_analysyis/OSA_and_BMI_exposure_to_T2D_outcome_BMI_UNadj_GWASs/SNPs_filtered_by_p/",threshold,"/Exposure_BMI_distinct_SNPs_filtered_by_p_",threshold,".csv"), row.names = FALSE)
